removed database helper from app.rb and moved into database_connection.rb
