+++ lib/database_connection.rb
@@ -1,0 +1,17 @@
+require 'pg'
+
+class DatabaseConnection
+
+  def self.connect(dbname)
+    @connection = PG.connect({hostname: '127.0.0.1', dbname: dbname})
+  end 
+
+  def self.exec_params(query, params)
+    @connection.exec_params(query, params)
+  end 
+
+  def self.exec(query)
+    @connection.exec(query)
+  end 
+
+end
